server/requirements.txt
flask
flask-smorest
python-dotenv
black
sqlalchemy
flask-sqlalchemy
flask-jwt-extended
passlib
flask-cors
flask-socketio
flask_migrate